SpectralPreprocessing now works for spectral regions fixing bug: can not paste into a property column
--HG--
extra : convert_revision : svn%3Ad1681fee-8e23-fc4c-b001-5eb8f207e95b/trunk%40296

+++ Altaxo/Base/Calc/Regression/PLS/PLSContentMemento.cs
@@ -131,7 +131,7 @@ namespace Altaxo.Calc.Regression.PLS
         // new in version 1
         if(info.CurrentElementName=="SpectralPreprocessingRegions")
         {
-          double[] regions;
+          int[] regions;
           info.GetArray("SpectralPreprocessingRegions", out regions);
           s._spectralPreprocessing.Regions = regions;
           s._spectralPreprocessing.Method  = (SpectralPreprocessingMethod)info.GetEnum("SpectralPreprocessingMethod",typeof(SpectralPreprocessingMethod));
